Added comments and readme
The comments are to show my  understanding of my code and the timer code that I used W3 schools to understand and implement. I also had some help from my friend who is a developer. (This help was only to explain to me what is going on with the timer because it is complex to me.)